#82 db: split schema migrations into per-file shape with runner + ledger
