+++ packages/server-for-chrome-extension/src/db/schema-migrations/README.md
@@ -1,0 +1,3 @@
+# Schema Migrations
+
+Each migration file is named `NNN-kebab-description.js` where `NNN` is a 3-digit zero-padded sequence number (e.g. `001-`, `002-`). The runner applies files in lexical sort order, which matches numerical order up to 999 — past 999 the sort breaks, so stay within that range. Each file exports `{ id, description, up(db, opts) }`. The `up()` function runs inside a transaction managed by the runner; do not open your own outer transaction unless you have a specific reason (the `001` migration uses a nested savepoint for the CHECK rewrite — that is fine). Write `up()` to be idempotent in spirit even though the ledger (`schema_migrations` table) guarantees single execution: restoring a vintage backup re-runs every migration cleanly only if `up()` short-circuits on already-applied state. Use current-state language in comments — no changelog narrative.
